requirements.txt
asgiref==3.5.2
black==22.12.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
coverage==6.5.0
Django==4.1.4
djangorestframework==3.14.0
Faker==15.3.4
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.13.1
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.23.5
pathspec==0.10.2
platformdirs==2.6.0
psutil==5.9.4
pycountry==22.3.5
pyenchant==3.2.2
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.2
